requirements.txt
dbt-bigquery
flask
google-cloud-logging